Inline command suggestion helper
